Addressing CR feedback for Helper methods implementing ImportantForAccessibility property.
